requirements.txt
Flask
flask-cors
Pillow
numpy
tf-nightly
requests
